+++ jxls-site/src/site/markdown/changes.md
@@ -6,6 +6,7 @@ v2.12.0
 * [#164 Update commons-compress](https://github.com/jxlsteam/jxls/issues/164)
 * [#147 Row height bugfix](https://github.com/jxlsteam/jxls/issues/147), contribution by [jools-uk](https://github.com/jools-uk)
 * [#153 Issue in Excel Output while using SXSSF Transformer](https://github.com/jxlsteam/jxls/issues/153)
+* [#129 GroupSum with condition](https://github.com/jxlsteam/jxls/issues/129) see example: GroupSumTest
 * [#73 Exceptions not propagated in EachCommand](https://github.com/jxlsteam/jxls/issues/73)
 
 With change #73 it's possible to throw exceptions instead of logging them as ERROR: